requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
certifi==2024.2.2
click==8.1.7
cmdstanpy==1.2.1
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.8
debugpy==1.8.0
decorator==5.1.1
dnspython==2.5.0
email-validator==2.1.0.post1
executing==2.0.1
fastapi==0.109.2
fonttools==4.48.1
h11==0.14.0
holidays==0.42
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
importlib-resources==6.1.1
iniconfig==2.0.0
ipykernel==6.29.2
ipython==8.21.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
orjson==3.9.13
packaging==23.2
pandas==2.2.0
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
plotly==5.17.0
pluggy==1.4.0
prompt-toolkit==3.0.43
prophet==1.1.5
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.0
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.2
Pygments==2.17.2
pyparsing==3.1.1
pytest==8.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
stanio==0.3.0
starlette==0.36.3
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
ujson==5.9.0
uvicorn==0.27.0.post1
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
scikit-learn==1.5.1

# Added dependencies
fastapi
uvicorn
tensorflow
streamlit

# matplotlib
# pandas
# pytest
# numpy
# prophet
# joblib
# plotly
# python-multipart
